Refactor dashboard and add components
